chore: refactor code for connectrpc, don't add connect GET options when the endpoints are google.api.http
