+++ internal/converter/testdata/standard/output/googleapi.openapi.json
@@ -246,35 +246,6 @@
         "title": "Foo2Response",
         "additionalProperties": false
       },
-      "encoding": {
-        "title": "encoding",
-        "enum": [
-          "proto",
-          "json"
-        ],
-        "description": "Define which encoding or 'Message-Codec' to use"
-      },
-      "base64": {
-        "type": "boolean",
-        "title": "base64",
-        "description": "Specifies if the message query param is base64 encoded, which may be required for binary data"
-      },
-      "compression": {
-        "title": "compression",
-        "enum": [
-          "identity",
-          "gzip",
-          "br"
-        ],
-        "description": "Which compression algorithm to use for this request"
-      },
-      "connect": {
-        "title": "connect",
-        "enum": [
-          "v1"
-        ],
-        "description": "Define the version of the Connect protocol"
-      },
       "google.protobuf.Empty": {
         "type": "object",
         "description": "A generic empty message that you can re-use to avoid defining duplicated\n empty messages in your APIs. A typical example is to use it as the request\n or the response type of an API method. For instance:\n\n     service Foo {\n       rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n     }"
