+++ internal/converter/testdata/standard/output/helloworld.openapi.json
@@ -239,6 +239,35 @@
         "additionalProperties": false,
         "description": "The response message containing the greetings"
       },
+      "encoding": {
+        "title": "encoding",
+        "enum": [
+          "proto",
+          "json"
+        ],
+        "description": "Define which encoding or 'Message-Codec' to use"
+      },
+      "base64": {
+        "type": "boolean",
+        "title": "base64",
+        "description": "Specifies if the message query param is base64 encoded, which may be required for binary data"
+      },
+      "compression": {
+        "title": "compression",
+        "enum": [
+          "identity",
+          "gzip",
+          "br"
+        ],
+        "description": "Which compression algorithm to use for this request"
+      },
+      "connect": {
+        "title": "connect",
+        "enum": [
+          "v1"
+        ],
+        "description": "Define the version of the Connect protocol"
+      },
       "connect-protocol-version": {
         "type": "number",
         "title": "Connect-Protocol-Version",
@@ -327,35 +356,6 @@
         },
         "additionalProperties": true,
         "description": "Contains an arbitrary serialized message along with a @type that describes the type of the serialized message, with an additional debug field for ConnectRPC error details."
-      },
-      "encoding": {
-        "title": "encoding",
-        "enum": [
-          "proto",
-          "json"
-        ],
-        "description": "Define which encoding or 'Message-Codec' to use"
-      },
-      "base64": {
-        "type": "boolean",
-        "title": "base64",
-        "description": "Specifies if the message query param is base64 encoded, which may be required for binary data"
-      },
-      "compression": {
-        "title": "compression",
-        "enum": [
-          "identity",
-          "gzip",
-          "br"
-        ],
-        "description": "Which compression algorithm to use for this request"
-      },
-      "connect": {
-        "title": "connect",
-        "enum": [
-          "v1"
-        ],
-        "description": "Define the version of the Connect protocol"
       }
     }
   },
