+++ internal/converter/testdata/standard/output/protovalidate-and-googleapi.openapi.yaml
@@ -56,28 +56,6 @@ components:
       required:
         - name
       additionalProperties: false
-    encoding:
-      title: encoding
-      enum:
-        - proto
-        - json
-      description: Define which encoding or 'Message-Codec' to use
-    base64:
-      type: boolean
-      title: base64
-      description: Specifies if the message query param is base64 encoded, which may be required for binary data
-    compression:
-      title: compression
-      enum:
-        - identity
-        - gzip
-        - br
-      description: Which compression algorithm to use for this request
-    connect:
-      title: connect
-      enum:
-        - v1
-      description: Define the version of the Connect protocol
 security: []
 tags:
   - name: foo.KnowledgeAssessmentReports
